bet365-api-scraper-main/requirements.txt
curl_cffi
python-dotenv
requests
urllib3
